Add closure, copy_closure, and copy_add_generators
The new methods replace the old copy and add generators constructor
(which is now private).

The method add_generators is also modified to take a vector of Element
pointers instead of an unordered_set of Element pointers, and the method
now adds everything from the vector as a generator of the semigroup,
regardless of whether there are duplicates or if the generators are
equal to existing elements of the semigroup.