elements: Remove incorrect assert
It can be that elements are defined over semirings that are distinct in
memory but equal (for example, when one element comes from a semigroup
and another from an ideal of that semigroup).  The previous assert
checkd if the pointers were equal, which would sometimes fail, when it
should not. So the assert is removed.