semigroups: fix incorrect copying of identity
This is in the private constructor used to partially copy a semigroup before
applying add_generators or closure. The problem occurred because previously
we were really_copy-ing the _id data member from the old semigroup into the
new, and if the degree changed and this was a semigroup of partial perms,
then _id->really_copy(deg_plus) did not give the identity on the new degree,
but rather was undefined on every point higher than the old degree.