semigroups: add factorisation method for RWSE
This is to facilitate changes to the Congruence method
nontrivial_classes. The method factorisation is renamed
minimal_factorisation (i.e. the factorisation returned by the
Froidure-Pin algorithm), and a new method factorisation is introduced
that simply calls minimal_factorisation in all previous cases. In the
case of an RWSE factorisation simply returns the word that is wrapped by
the RWSE, which may not be minimal, hence the name change for the
methods.