cong: fix issue with nontrivial_classes method
The new KBP created in the Congruence::nontrivial_classes method
was not properly deleted. Fixing this highlighted a problem where
Partition's were copied (when returned from functions, and assigned
to variables), then when the original (which was not a pointer) went out
of scope it was destructed, and hence the copy was not any longer valid
(since it only copied the pointers to the actual data). To avoid this, this
commit deletes the copy constructor for a Partition, and makes
every nontrivial_classes method return a pointer.