cong: fix Heisenbug in Congruence::DATA method
This showed up in the travis tests (on the release of 0.2.0) and
was in Congruence 05. It could be repeated by running the test
in Congruence 05 about 100000 times. The issue was that in
some circumstances 1 thread completed (is_done) and was
subsequently killed by and killed another thread. This meant that
every thread was killed, but the methods depending on this used
is_killed to determine whether a thread had won or if it could
an answer. This commit checks for is_done and/or is_killed, and
resolves the issue.