tc: in init, only set _tc_done to true if not killed
If a TC object was killed while prefilling, it might have empty
relations during init().  But this doesn't mean it's done.  We should
only allow it to set _tc_done to true if it hasn't already been killed.

This issue was revealed by commit 668d6ea.

I also removed a FIXME: even if we are prefilled here, we still always
need to apply the "extra" relations.