semigroups: fix limit overflow if _batch_size is very large
This was highlighted by the KBFP::run method being called with
Congruence::LIMIT_MAX steps, where the limit is taken to be
std::max(limit, _nr + _batch_size), the initial value of limit is small,
and _nr + _batch_size overflows and so the limit was set to a very
small value (2). This commit checks that _nr + _batch_size does
not overflow, before taking the std::max, and if it does overflow,
we use the _batch_size for the limit instead.