build: use rpmalloc instead of malloc
This commit makes libsemigroups use the memory allocator rpmalloc
instead of the system malloc. The reason for this is that rpmalloc
improves the performance of libsemigroups. The version of rpmalloc used
is 1.2, it is slightly modified to define _GNU_SOURCE if it is not
already defined, and to do nothing if the preprocessor symbol __CYGWIN__
or __CYGWIN32__ is defined.  Defining _GNU_SOURCE is required to use
rpmalloc as a drop-in replacement for malloc when compiling with gcc.
rpmalloc also uses some things that are not implemented in cygwin, and
so it is disabled under cygwin.

Using rpmalloc can be disabled by configuring with the option
--enable-system-malloc.