semigroups: implement iterators and const
This commit replaces all accesses to the _elements data member of
Semigroup with Element const* instead of Element*, and removes the
elements() methods and replaces it with const_iterator, and methods for
cbegin, cend. This is somewhat more complicated than required at
present, but this is intended to make changing to the multithreaded
Froidure-Pin easier. There are a number of changes in the tests etc to
accommodate these changes.