elements: don't use thread_local with GCC <= 4.9
There is a compiler error in GCC 4.8.? which cases an internal error
when compiling code with templated thread_local static storage, this
commit disables use of thread_local when the using any version of GCC
before 4.9.  This will hopefully resolve an issue with the last attempt
to update the conda package.