semigroups: add SemigroupBase class
The purpose of this is to make it easier to write binding for
libsemigroups in other languages, principally GAP. This change allows
access to the methods of a Semigroup<TElementType> class that do not
have any parameters or return value involving TElementType objects, so
that a single function can access such methods via a pointer to a
SemigroupBase (i.e. no duplicate functions just to call the method
Semigroup<TElementType1>->size() and Semigroup<TElementType2>->size()
and so on).