cong-intf: use lazy FroidurePin
The purpose of this is to allow us to store an FpSemigroup inside any
CongruenceInterface object and then to only compute the FroidurePin
instance, representing the semigroup over which the CongruenceInterface
is defined, at the point when it is required. At present this point is
when the non-trivial classes are computed. The computation of the
FroidurePin instance might be infinite, and so we don't want to trigger
it unless absolutely necessary. If it is infinite, then either there are
infinitely many non-trivial classes or at least one of those classes is
infinite, in which case the non-trivial class computation will be
infinite also (so nothing is lost by this approach). On the other hand,
the CongruenceInterface object might represent a finite congruence over
an infinite semigroup, and we can then compute happily with the
CongruenceInterface, without triggering the computation of its parent
FroidurePin object, unless the (anyway infinite) computation of the
non-trivial classes is requested.