felsch-tree: refactor to allow free generators
When FelschTree was inside ToddCoxeter it would never have been the case that
there were free generators when the FelschTree was constructed because
ToddCoxeter refuses to run when it is obviously infinite. Now, when using the
FelschTree in Sims1 the situation is different, and we can legitimately have the
where there are free generators, the changes in this PR reflect allow this.