feat: prune and compact rules list
The top level for-loop now checks to see if the number of removed rules
has reached a certain threshold. If it has, then the zombie entries are
removed from the rules and active vectors, and the remaining active
rules are removed and readded to the lookup in case their indices have
changed. The change requires benchmarking to determine both the
threshold, and if the implementation is sufficiently optimised.