Fix issue with fmt-v11.1.4
The issue is that our custom formatters for POSITIVE_INFINITY +
NEGATIVE_INFINITY no longer work. I tried to fix this "properly"
but couldn't figure out if it is an us problem or a me problem.

I also incorporated better exception messages including the integer
values of infinity where appropriate in matrix.hpp.