todd-coxeter: fix bug in reduce_no_run_no_checks
The previous code was invalid for onesided congruences with generating
pairs previously, because we rewrote all possible subwords of the input
word, which in a onesided (right) congruence it's only valid to rewrite
prefixes. The only rewriting we can do on prefixes is to follow the
least path from the initial node to the target node of the path labelled
by the input word, because no prefix can be rewritten or otherwise the
least path wouldn't be the least path.