todd-coxeter: improve reporting
This involved extensive refactoring, including swapping the order of
NodeManagedGraph and FelschGraph in the class hierarchy; added edge
counting and stats accessors, and improved edge registration. A fair
amount of redundant or comment out code and tests was also removed.
Reporting was moved out of NodeManagedGraph and into ToddCoxeter and
Stephen.